+++ src/DotRecast.Detour.Crowd/DtCrowdTelemetry.cs
@@ -27,24 +27,6 @@ namespace DotRecast.Detour.Crowd
 {
     public class DtCrowdTelemetry
     {
-        public readonly struct DisposableHandle : IDisposable
-        {
-            private readonly DtCrowdTimerLabel _label;
-            private readonly DtCrowdTelemetry _telemetry;
-
-            public DisposableHandle(DtCrowdTelemetry telemetry, DtCrowdTimerLabel label)
-            {
-                _telemetry = telemetry;
-                _label = label;
-            }
-
-            public void Dispose()
-            {
-                _telemetry.Stop(_label);
-            }
-        }
-            
-        
         public const int TIMING_SAMPLES = 10;
         private float _maxTimeToEnqueueRequest;
         private float _maxTimeToFindPath;
@@ -87,18 +69,17 @@ namespace DotRecast.Detour.Crowd
             _maxTimeToFindPath = Math.Max(_maxTimeToFindPath, time);
         }
 
-        public DisposableHandle ScopedTimer(DtCrowdTimerLabel label)
+        internal DtCrowdScopedTimer ScopedTimer(DtCrowdTimerLabel label)
         {
-            Start(label);
-            return new DisposableHandle(this, label);
+            return new DtCrowdScopedTimer(this, label);
         }
 
-        private void Start(DtCrowdTimerLabel name)
+        internal void Start(DtCrowdTimerLabel name)
         {
             _executionTimings.Add(name, RcFrequency.Ticks);
         }
 
-        private void Stop(DtCrowdTimerLabel name)
+        internal void Stop(DtCrowdTimerLabel name)
         {
             long duration = RcFrequency.Ticks - _executionTimings[name];
             if (!_executionTimingSamples.TryGetValue(name, out var cb))
@@ -110,7 +91,5 @@ namespace DotRecast.Detour.Crowd
             cb.PushBack(duration);
             _executionTimings[name] = (long)cb.Average();
         }
-
-
     }
 }